Execute listener tests
Without the suffix "Test.php" in the file names, they were being
ignored. Some fixes and cleanup were made but the majority of the tests
do not rely on real responses from the Trello API, which needs to be
fixed on a next occasion - therefore the @TODOs added.
