+++ lib/Trello/Exception/ApiLimitExceedException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Trello\Exception;
+
+class ApiLimitExceedException extends \InvalidArgumentException implements ExceptionInterface
+{
+
+}
